refactor modal components
