--- src/components/modal/modal-template/style.module.scss
@@ -1,13 +1,0 @@
-.modal {
-  padding: 0;
-  width: min(37.5rem, 100%);
-  min-height: 18.75rem;
-  border: none;
-  border-radius: 0.875rem;
-  bottom: auto;
-  top: 7.25vmin;
-
-  &::backdrop {
-    background-color: rgba(0, 0, 0, 0.4);
-  }
-}
