+++ src/components/modal/reply-tweet-modal/index.tsx
@@ -1,22 +1,25 @@
 // import { tweetsData } from "../index"
 import Image from "next/image"
 import fakePhoto from "../../../assets/img/fake-photo.png"
-import classes from "./style.module.scss"
-import PostTweetModal, { PostTweetModalProps } from "../post-tweet-modal"
+import classes from "../style.module.scss"
+import { ModalPostProps } from "../modal-post"
 import Link from "next/link"
-import modalClasses from "../style.module.scss"
+import ModalTemplate from "../modal-template"
+import Header from "components/header"
+import { useRef } from "react"
+import Button from "components/button"
 
-interface ReplyTweetModalProps extends PostTweetModalProps {
+interface ReplyTweetModalProps extends ModalPostProps {
   replyTweetInfo: {
-    tweetId: string | number
-    content: string | number
+    readonly tweetId: string | number
+    readonly content: string | number
     owner: {
-      id: string
-      name: string
-      account: string
-      avatar?: string
+      readonly id: string | number
+      readonly name: string
+      readonly account: string
+      readonly avatar?: string
     }
-    createdAt: number
+    readonly createdAt: number
   }
 }
 
@@ -27,27 +30,42 @@ const ReplyTweetModal = ({
   postApi,
   replyTweetInfo
 }: ReplyTweetModalProps) => {
+  const textareaRef = useRef<HTMLTextAreaElement>(null)
   const {
     tweetId,
     owner: { id, name, account, avatar },
     content,
     createdAt
   } = replyTweetInfo
+  const handleDialogClose = () => {
+    const textarea = textareaRef.current
+    const textContent = textarea?.value
+    if (!textContent) return
+    textarea.value = ""
+  }
+  const handleFormSubmit = () => {
+    const textarea = textareaRef.current
+    const textContent = textarea?.value
+    if (!textContent) return
+    // TODO 傳送新的推文
+    console.log(textarea.value)
+    postApi()
+    textarea.value = ""
+  }
 
   return (
-    <PostTweetModal
-      isVisible={isVisible}
-      onDialogClose={onDialogClose}
-      currentUser={currentUser}
-      postApi={postApi}>
-      <aside className={`${modalClasses.border_top} ${classes["reply-modal"]}`}>
-        <div className={modalClasses.avatar}>
+    <ModalTemplate isVisible={isVisible} onDialogClose={handleDialogClose}>
+      <Header handleLeftClick={onDialogClose} utility="modal" />
+      <aside className={classes.modal__reply}>
+        <div className={classes.avatar}>
           <Link href={`/${id}`}>
-            <Image src={avatar || fakePhoto} alt="tweet owner's avatar" />
+            <>
+              <Image src={avatar || fakePhoto} alt="tweet owner's avatar" />
+            </>
           </Link>
           <div className={classes.decoration}></div>
         </div>
-        <section className={classes["reply-modal__content"]}>
+        <section className={classes.modal__content}>
           <p>
             <strong>
               <Link href={`/${id}`}>{name}</Link>{" "}
@@ -65,7 +83,32 @@ const ReplyTweetModal = ({
           </p>
         </section>
       </aside>
-    </PostTweetModal>
+      <aside className={classes.modal__post}>
+        <div className={classes.avatar}>
+          <Link href={`/${currentUser.id}`}>
+            <>
+              <Image
+                src={currentUser.avatar || fakePhoto}
+                alt="current user's avatar"
+              />
+            </>
+          </Link>
+        </div>
+        <form
+          method="dialog"
+          onSubmit={handleFormSubmit}
+          className={classes.form}>
+          <textarea
+            ref={textareaRef}
+            placeholder="有什麼新鮮事？"
+            className={classes.textarea}
+          />
+          <Button className={classes.form__btn}>
+            <span>推文</span>
+          </Button>
+        </form>
+      </aside>
+    </ModalTemplate>
   )
 }
