--- src/components/modal/reply-tweet-modal/style.module.scss
@@ -1,31 +1,0 @@
-.reply-modal {
-  padding: 1.125rem 0.9375rem 0.625rem;
-  gap: 0.625rem;
-  font-size: var(--fz-300);
-  display: grid;
-  grid-template-columns: auto 1fr;
-
-  &__content {
-    width: min(100%, 31.875rem);
-    word-wrap: break-word;
-    & p:first-child {
-      margin-block-end: 0.3125rem;
-    }
-
-    & p:last-child {
-      margin-block-start: var(--space-4);
-      font-size: var(--fz-200);
-
-      & a {
-        color: var(--clr-primary);
-      }
-    }
-  }
-}
-
-.decoration {
-  height: calc(100% - 1.5626rem - 0.625rem);
-  background-color: var(--clr-border);
-  width: 1px;
-  translate: 1.5rem;
-}
